// src/app/model/IMeme.model.ts
/**
 * @description This interface represents model structur of a Meme 
 * @property {number} id Unique autogenerated ID of Meme 
 * @property {string} name Name of Uploader of Meme 
 * @property {string} caption Caption of Meme 
 * @property {string} url URL to Image used in Meme
 */
export interface IMeme {
    id: number,
    name: string,
    caption: string,
    url: string
}